deep_research_slides/backend/pyproject.toml
[project]
name = "deep-research-slides-backend"
version = "0.1.0"
description = "Deep Research to Slides Generator Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.7",
    "anthropic>=0.37.1",
    "beautifulsoup4>=4.12.3",
    "datasets>=2.21.0",
    "google-search-results>=2.4.2", 
    "huggingface-hub>=0.23.4",
    "markdownify>=0.13.1",
    "openai>=1.52.2",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "transformers>=4.46.0",
    "jinja2>=3.1.2",
    "markdown>=3.5.0",
    "litellm>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]